requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
pydantic==2.9.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
yookassa==3.3.0
httpx==0.27.0
